refactor: 组件名称格式统一
