+++ app/assets/components/bulk-editor/index.jsx
