+++ app/assets/components/editor/index.less
