+++ app/assets/components/global-search/index.jsx
@@ -2,7 +2,7 @@ import React from 'react';
 import { Icon, Input, AutoComplete } from 'antd';
 import ajax from 'xhr-plus';
 
-import './GlobalSearch.less';
+import './index.less';
 
 const Option = AutoComplete.Option;
 const OptGroup = AutoComplete.OptGroup;
