+++ app/assets/components/global-search/index.less
