+++ app/assets/components/info/index.less
