+++ app/assets/components/schema-editor/index.less
