+++ app/assets/pages/api-mock-page/index.jsx
@@ -6,8 +6,8 @@ import Mock from 'mockjs';
 import CopyToClipboard from 'react-copy-to-clipboard';
 import hljs from 'highlight.js/lib/highlight';
 
-import Info from '../../components/Info';
-import Editor from '../../components/Editor';
+import Info from '../../components/info/index';
+import Editor from '../../components/editor/index';
 
 import './index.less';
 import 'highlight.js/styles/googlecode.css';
